Bugfix - zoom level should be a float to fit the crop factor and focal lenght resizing

+++ src/main/java/com/threeamigos/imageviewer/implementations/datamodel/ImageSliceImpl.java
@@ -105,7 +105,7 @@ public class ImageSliceImpl implements ImageSlice, PropertyChangeListener {
 	}
 
 	@Override
-	public void changeZoomLevel(int zoomLevel) {
+	public void changeZoomLevel(float zoomLevel) {
 		pictureData.changeZoomLevel(zoomLevel);
 	}
